Always enable the tests
Instead of overriding the meaning of static, allow explicitly exposing
methods to tests with the "exposed_to_tests" define. When this is used,
only while compiling tests will the methods be non-static. Methods
exposed to tests must be placed in the *_tests.h headers which are
included from test files.