Flesh out the class hierarchy of the server
This is the first commit that tries to establish a class hierarchy for
the server. Now the caching server implementations act like they are the
implementations of the virtual methods on the command_buffer_server_t
class.

The tests now call these functions as if they are methods as well.