Make glDrawElements/glDrawArrays more effeccient
Previously to determine whether or not the vertex attribute data was
small enough to fit into one array, the code would loop over every
vertex attribute. Instead we record the attribute data (and not just the
pointer) of the first and last attribute pointer. This avoids having to
loop through all attributes.