Simplify the name_handler
We only need to store one set of names. On the client it doesn't matter
whether the names of resources exist in separate namespaces. All that
matters is that each name is unique and we can map it on the server to
the real name. We never have to map from the real name to the fake name.