Reuse old name_handler names more readily
This fixes a memory leak due to the fact that the hash table
implementation does no rehashing. Thus if we continue to add names with
larger and larger value, the hash table will continue to grow in size.

We work around this issue by simply saving a list of freed names and
not using hash table entirely. This should both speed up the operation
of the name handler and use less memory.