implement checkpoint egl/gl call order list.
A checkpoint is an indication of the order of gl/egl calls
from different threads (in the future from different processes).
The server keeps a queue of checkpoints.  The client generates
a checkpoint log message when it encounters egl/gl calls that
must be kept in the order.  These include create display/surface/
context, destroy display/surface/context, make current, any
flush/finish, (un)binding surface to image/texture.

When a client encounters such calls, it first send a log message to
the server.  The server then puts it the queue.  The client then
sends the the call through command buffer.  When the server
receives from the command buffer, it checks whether this command
matches the head of the queue.  If it does not match, the server waits,
otherwise, it removes the queue head and proceeds.